Backend/creditai-backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation / Settings
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# AI
google-generativeai==0.8.3

# ML
xgboost==2.1.3
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2

# PDF generation
reportlab==4.2.5

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
